requirements.txt
prometheus-client
requests>=2.27.1
pyyaml
